Added support for dividend through bank statement.

+++ django_gotolong/trendlyne/models.py
@@ -6,7 +6,7 @@ from django.db import models
 class Trendlyne(models.Model):
     tl_stock_name = models.TextField(blank=True, null=True)
     tl_isin = models.TextField(primary_key=True)
-    tl_bat = models.IntegerField(blank=True, null=True)
+    tl_bat = models.FloatField(blank=True, null=True)
     tl_der = models.FloatField(blank=True, null=True)
     tl_roce3 = models.FloatField(blank=True, null=True)
     tl_roe3 = models.FloatField(blank=True, null=True)
